test: HTTP 401 logout test
